move deepMerge to helpers
